isolate demo element and used element utils
