+++ BUILD.bazel
@@ -14,6 +14,7 @@ cc_library(
         name = src[:-len(".cpp")],
         size = "small",
         srcs = [src],
+        hdrs = glob(["sfo_cpp/tests/*.hpp"]),
         args = ["--gtest_output"],
         copts = ["-std=c++17"],  # un-comment for *nix
         # copts = ["/std:c++17"],  # un-comment for windows
